requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
streamlit==1.27.0
streamlit-geolocation==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.13.0
transformers==4.35.0
torch==2.2.2
pydantic==1.10.11
phonenumbers==8.13.17
python-multipart==0.0.6
requests==2.31.0
dotnev